+++ index.d.ts
@@ -29,6 +29,12 @@ declare module '@amaas/amaas-core-sdk-js' {
       function retire({ AMId, resourceId }: { AMId: number, resourceId: string }, callback?: Function): Promise<books.Book> | void
       function reactivate({ AMId, resourceId }: { AMId: number, resourceId: string }, callback?: Function): Promise<books.Book> | void
     }
+    namespace Fundamentals {
+      function countries({ code }: { code: string }, callback?: Function): Promise<{ [key: string]: string | number }> | void
+      function calcBusinessDate({ date, codes, offset, invalidDates }: { date: string, codes: string | string[], offset: number, invalidDates: any }, callback?: Function): Promise<{ businessDate: string }> | void
+      function processDateInfo({ date, codes }: { date: string, codes: string | string[] }, callback?: Function): Promise<{ [countryCode: string]: any }> | void
+      function holidays({ codes, years }: { codes: string | string[], years: string | string[] }, callback?: Function): Promise<{ [countryCode: string]: { calendar: string, holidays: any[], weekends: number[] } }> | void
+    }
     namespace Monitor {
       function retrieveItems({ AMId, resourceId }: { AMId: number, resourceId?: string }, callback?: Function): Promise<monitor.Item | monitor.Item[]> | void
       function insertNewItem({ AMId, item }: { AMId: number, item: monitor.IMonitorItem | monitor.Item }, callback?: Function): Promise<monitor.Item> | void