+++ src/utils/fundamentals/fundamentals.test.js
@@ -61,25 +61,25 @@ describe('utils/fundamentals', () => {
         })           
         it('calls searchData with the correct parameters', done => {
               fundamentals.calcBusinessDate( {date: '2017-6-30', codes: 'SG', offset: 1}, (error, result) => {
-              expect(network.searchData).toHaveBeenCalledWith({ AMaaSClass: 'fundamentalBusinessDate', query: {startDate: ['2017-6-30'], countryCode: ['SG'], offset: [1], invalidDate: []} })
+              expect(network.searchData).toHaveBeenCalledWith({ AMaaSClass: 'fundamentalBusinessDate', query: {startDate: ['2017-6-30'], countryCodes: ['SG'], offset: [1], invalidDates: []} })
               done()
             })
         })
         it('calls searchData with the correct parameters', done => {
               fundamentals.calcBusinessDate( {date: '2017-6-30', codes: ['SG', 'CHN'], offset: 1}, (error, result) => {
-              expect(network.searchData).toHaveBeenCalledWith({ AMaaSClass: 'fundamentalBusinessDate', query: {startDate: ['2017-6-30'], countryCode: ['SG','CHN'], offset: [1], invalidDate: []} })
+              expect(network.searchData).toHaveBeenCalledWith({ AMaaSClass: 'fundamentalBusinessDate', query: {startDate: ['2017-6-30'], countryCodes: ['SG','CHN'], offset: [1], invalidDates: []} })
               done()
             })
         })
         it('calls searchData with the correct parameters', done => {
               fundamentals.calcBusinessDate( {date: '2017-6-30', codes: 'SG', offset: 1, invalidDates: '2017-7-01'}, (error, result) => {
-              expect(network.searchData).toHaveBeenCalledWith({ AMaaSClass: 'fundamentalBusinessDate', query: {startDate: ['2017-6-30'], countryCode: ['SG'], offset: [1], invalidDate: ['2017-7-01']} })
+              expect(network.searchData).toHaveBeenCalledWith({ AMaaSClass: 'fundamentalBusinessDate', query: {startDate: ['2017-6-30'], countryCodes: ['SG'], offset: [1], invalidDates: ['2017-7-01']} })
               done()
             })
         })
         it('calls searchData with the correct parameters', done => {
               fundamentals.calcBusinessDate( {date: '2017-6-30', codes: 'SG', offset: 1, invalidDates: ['2017-7-01', '2017-7-02']}, (error, result) => {
-              expect(network.searchData).toHaveBeenCalledWith({ AMaaSClass: 'fundamentalBusinessDate', query: {startDate: ['2017-6-30'], countryCode: ['SG'], offset: [1], invalidDate: ['2017-7-01','2017-7-02']} })
+              expect(network.searchData).toHaveBeenCalledWith({ AMaaSClass: 'fundamentalBusinessDate', query: {startDate: ['2017-6-30'], countryCodes: ['SG'], offset: [1], invalidDates: ['2017-7-01','2017-7-02']} })
               done()
             })
         })
@@ -95,13 +95,13 @@ describe('utils/fundamentals', () => {
         })           
         it('calls searchData with the correct parameters', done => {
               fundamentals.processDateInfo( {date: '2017-7-03', codes: 'SG'}, (error, result) => {
-              expect(network.searchData).toHaveBeenCalledWith({ AMaaSClass: 'fundamentalDateInfo', query: {startDate: ['2017-7-03'], countryCode: ['SG']} })
+              expect(network.searchData).toHaveBeenCalledWith({ AMaaSClass: 'fundamentalDateInfo', AMId: '2017-7-03', query: { countryCodes: ['SG'] } })
               done()
             })
         })
         it('calls searchData with the correct parameters', done => {
               fundamentals.processDateInfo( {date: '2017-7-03', codes: ['SG', 'CHN']}, (error, result) => {
-              expect(network.searchData).toHaveBeenCalledWith({ AMaaSClass: 'fundamentalDateInfo', query: {startDate: ['2017-7-03'], countryCode: ['SG', 'CHN']} })
+              expect(network.searchData).toHaveBeenCalledWith({ AMaaSClass: 'fundamentalDateInfo', AMId: '2017-7-03', query: { countryCodes: ['SG', 'CHN'] } })
               done()
             })
         })
@@ -117,19 +117,19 @@ describe('utils/fundamentals', () => {
         })
         it('calls searchData with the correct parameters', done => {
               fundamentals.holidays( {codes: 'SG', years: '2017'}, (error, result) => {
-              expect(network.searchData).toHaveBeenCalledWith({ AMaaSClass: 'fundamentalHoliday', query: {countryCode: ['SG'], year: ['2017']} })
+              expect(network.searchData).toHaveBeenCalledWith({ AMaaSClass: 'fundamentalHoliday', query: { countryCodes: ['SG'], year: ['2017'] } })
               done()
             })
         })           
         it('calls searchData with the correct parameters', done => {
               fundamentals.holidays( {codes: ['SG','CHN'], years: '2017'}, (error, result) => {
-              expect(network.searchData).toHaveBeenCalledWith({ AMaaSClass: 'fundamentalHoliday', query: {countryCode: ['SG', 'CHN'], year: ['2017']} })
+              expect(network.searchData).toHaveBeenCalledWith({ AMaaSClass: 'fundamentalHoliday', query: { countryCodes: ['SG', 'CHN'], year: ['2017'] } })
               done()
             })
         })       
         it('calls searchData with the correct parameters', done => {
               fundamentals.holidays( {codes: 'SG', years: ['2017', '2018']}, (error, result) => {
-              expect(network.searchData).toHaveBeenCalledWith({ AMaaSClass: 'fundamentalHoliday', query: {countryCode: ['SG'], year: ['2017', '2018']} })
+              expect(network.searchData).toHaveBeenCalledWith({ AMaaSClass: 'fundamentalHoliday', query: {countryCodes: ['SG'], year: ['2017', '2018']} })
               done()
             })
         })