+++ src/utils/fundamentals/index.js
@@ -1,0 +1,7 @@
+/**
+ * @namespace Fundamentals
+ * @memberof module:api
+ */
+
+import { countries, calcBusinessDate, processDateInfo, holidays } from './fundamentals'
+export { countries, calcBusinessDate, processDateInfo, holidays }