+++ CHANGELOG.md
@@ -13,6 +13,7 @@ All notable changes to this project will be documented in this file.
 - Add monitor api calls (insertNewItem, resubmitItem, searchItems, closeItem).
 - Add domains sub-service to AssetManagers.
 - Add EODBook class and retrieveEODBooks function to AssetManagers module.
+- Add fundamentals service.
 
 ### Changed
 - Removed `maturityDate` from the Asset base class and set either maturityDate or expiryDate on the relevant subclasses.